+++ shop-the-look/.env.development.example
@@ -1,4 +1,4 @@
-GOOGLE_CREDENTIALS_BASE64=
-PINECONE_API_KEY=
+GOOGLE_CREDENTIALS_BASE64= # Instructions on retrieving base64-encoded Google Credentials: https://github.com/pinecone-io/sample-apps/blob/main/shop-the-look/README.md#setup-api-keys
+PINECONE_API_KEY= # Instructions on retrieving Pinecone API key: https://github.com/pinecone-io/sample-apps-internal/blob/main/shop-the-look/README.md#setup-api-keys
 NEXT_PUBLIC_DEVELOPMENT_URL=http://localhost:8000 # Change to your backend URL if needed. For production, Vercel will set the appropriate value, so you don't need this variable.
 NEXT_PUBLIC_VERCEL_ENV=demo # Change to "development" to use your own backend. For production, Vercel will set the appropriate value, so you don't need this variable.
