+++ shop-the-look/LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2023 Diego Valdez
+Copyright (c) 2024 Pinecone Systems, Inc
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
